issue #4 startup parameters moved from Parameters module to run module. Startup code converted to class
