+++ sinner/Parameters.py
@@ -1,40 +1,15 @@
-import platform
 import shlex
 import sys
 from argparse import ArgumentParser, Namespace
 from typing import List
 
-from sinner.validators.AttributeLoader import AttributeLoader, Rules
 
-
-class Parameters(AttributeLoader):
-    gui: bool
-    benchmark: bool
-    max_memory: int
-
+class Parameters:
     parser: ArgumentParser = ArgumentParser()
     parameters: Namespace
 
-    def rules(self) -> Rules:
-        return [
-            {
-                'parameter': 'max-memory',
-                'default': self.suggest_max_memory()
-            },
-            {
-                'parameter': 'gui',
-                'default': False
-            },
-            {
-                'parameter': 'benchmark',
-                'default': None,
-            },
-        ]
-
     def __init__(self, command_line: str | None = None):
         self.parameters = self.command_line_to_namespace(command_line)
-        super().__init__(self.parameters)
-        self.parameters.max_memory = self.max_memory  # add initialized value to use it later
 
     @staticmethod
     def command_line_to_namespace(cmd_params: str | None = None) -> Namespace:
@@ -78,9 +53,3 @@ class Parameters(AttributeLoader):
         else:  # '--key value1 value2'
             key, value = argument[2:].split('=')
             return key, value.split()
-
-    @staticmethod
-    def suggest_max_memory() -> int:
-        if platform.system().lower() == 'darwin':
-            return 4
-        return 16
